better the flipping experience, start the shooter selector

+++ src/main.cpp
@@ -1,3 +1,4 @@
+#include <concepts>
 import sfml;
 import fmt;
 import stdutils;
@@ -7,6 +8,7 @@ import stringables;
 import tilemap;
 import json;
 import ginseng;
+import random;
 
 void my_assert(bool condition,
   std::string_view message = "",
@@ -22,6 +24,14 @@ void my_assert(bool condition,
         throw std::runtime_error(msg);
     }
 }
+
+template<typename T>
+concept with_bounds =
+  std::convertible_to<T, sf::Transformable> && requires(T t) {
+      { t.getGlobalBounds() } -> std::convertible_to<sf::FloatRect>;
+      { t.getLocalBounds() } -> std::convertible_to<sf::FloatRect>;
+  };
+
 namespace components {
 
 struct Animation : public sf::Sprite
@@ -70,13 +80,15 @@ struct Velocity
     float magnitude;
 };
 
+using draw_bounding_box = ginseng::tag<struct draw_bounding_box_t>;
+
 struct BoundingBox : public sf::RectangleShape
 {
-    BoundingBox(sf::Sprite &sprite)
+    BoundingBox(const with_bounds auto &sprite)
     {
         setSize(
           { sprite.getGlobalBounds().width, sprite.getGlobalBounds().height });
-        setPosition(sprite.getPosition());
+        setPosition(sprite.getPosition() - sprite.getOrigin());
         setFillColor(sf::Color::Transparent);
         setOutlineColor(sf::Color::Red);
         setOutlineThickness(2);
@@ -101,12 +113,12 @@ using player_tag = ginseng::tag<struct player_tag_t>;
 
 using namespace components;
 
-const sf::Vector2u win_size{ 640, 640 };
+const sf::Vector2u win_size{ 640, 704 };
 
 constexpr auto font_path = "/System/Library/Fonts/Times.ttc";
 
 
-auto sprite_center(const sf::Sprite &sprite)
+auto sprite_center(const with_bounds auto &sprite)
 {
     return sf::Vector2f{ sprite.getGlobalBounds().width / 2,
         sprite.getGlobalBounds().height / 2 };
@@ -206,11 +218,18 @@ class EntityFactory
         auto frames_json = nlohmann::json::parse(anim_frames_file);
 
         demon_frames =
-          load_animation_frames_json(frames_json, "big_demon_idle_anim");
+          load_animation_frames_json(frames_json, "big_demon_idle");
         zombie_frames =
-          load_animation_frames_json(frames_json, "big_zombie_idle_anim");
-        wizard_frames =
-          load_animation_frames_json(frames_json, "wizzard_m_idle_anim");
+          load_animation_frames_json(frames_json, "big_zombie_idle");
+        wizard_frames = load_animation_frames_json(frames_json, "wizzard_idle");
+        {
+            my_assert(frames_json.contains("wizzard_idle_origin")
+                      && frames_json["wizzard_idle_origin"].is_array());
+            std::vector<float> vecarr = frames_json["wizzard_idle_origin"];
+            wizard_origin.x = vecarr[0];
+            wizard_origin.y = vecarr[1];
+        }
+
         texture = manager.get<sf::Texture>("../characters.png");
     }
 
@@ -247,14 +266,13 @@ class EntityFactory
         return id;
     }
 
-    auto spawn_wizard(ginseng::database &db)
+    auto spawn_wizard(ginseng::database &db, sf::Vector2f position = {})
     {
         auto id = db.create_entity();
         components::Animation anim{ &wizard_frames, *texture, 0.1f };
+        anim.setOrigin(wizard_origin);
         anim.setScale(2.f, 2.f);
-        anim.setPosition(
-          enemies_turning_points.front() - sf::Vector2f{ 0, 100 });
-        anim.setOrigin(sprite_center(anim));
+        anim.setPosition(position);
         db.add_component(id, anim);
 
         db.add_component(id, components::player_tag{});
@@ -263,6 +281,7 @@ class EntityFactory
 
     auto get_zombie_frames() const { return zombie_frames; }
     auto get_demon_frames() const { return demon_frames; }
+    auto get_wizard_frames() const { return wizard_frames; }
     auto get_enemy_turning_points() const { return enemies_turning_points; }
 
   private:
@@ -273,8 +292,87 @@ class EntityFactory
     std::vector<sf::IntRect> demon_frames;
     std::vector<sf::IntRect> wizard_frames;
     std::vector<sf::IntRect> zombie_frames;
+    sf::Vector2f wizard_origin;
     std::vector<sf::Vector2f> enemies_turning_points;
 };
+
+enum class shooter_type {
+    none = 0,
+    wizard,
+};
+
+struct shooter_select : public sf::Drawable
+{
+    shooter_select(sf::Color color,
+      shooter_type type,
+      const sf::Texture &text,
+      sf::IntRect rect)
+      : color(color), type(type)
+    {
+        my_assert(text.getSize().x > 0 && text.getSize().y > 0,
+          "texture size must be greater than 0");
+        sprite.setTexture(text);
+        sprite.setTextureRect(rect);
+        sprite.setPosition({ 0, 640 });
+        // scale to be 64x64
+        // auto cal_scale = 64.f / text.getSize().x;
+        sprite.setScale(2, 2);
+    }
+    void draw(sf::RenderTarget &target, sf::RenderStates states) const override
+    {
+        sf::RectangleShape shooter_selector_bg;
+        shooter_selector_bg.setSize({ 64, 64 });
+        shooter_selector_bg.setPosition({ 0, 640 });
+        shooter_selector_bg.setFillColor(color);
+        target.draw(shooter_selector_bg, states);
+        target.draw(sprite, states);
+    }
+    sf::Color color;
+    shooter_type type;
+    sf::Sprite sprite;
+};
+
+void flip_sprite(sf::Transformable &target, bool flip_x, bool flip_y)
+{
+    auto g = target.getScale();
+    g.x = std::abs(g.x);
+    g.y = std::abs(g.y);
+    target.setScale(flip_x ? -g.x : g.x, flip_y ? -g.y : g.y);
+}
+
+void draw_select_tile(sf::RenderTarget &target,
+  sf::Vector2f mouse_pos,
+  sf::Vector2u grid_space_size = { 32, 32 })
+{
+    sf::RectangleShape grid_tile;
+    grid_tile.setFillColor(sf::Color{ 0x32323255 });
+    grid_tile.setOutlineColor(sf::Color{ 0x111111ff });
+    grid_tile.setOutlineThickness(1);
+    grid_tile.setSize(sf::Vector2f{ grid_space_size });
+    const auto rows = win_size.y / std::size_t(grid_space_size.y);
+    const auto columns = win_size.x / std::size_t(grid_space_size.x);
+    for (std::size_t i = 0; i < rows; i++)
+    {
+        for (std::size_t j = 0; j < columns; j++)
+        {
+            grid_tile.setPosition(grid_space_size.x * i, grid_space_size.y * j);
+            if (grid_tile.getGlobalBounds().contains(mouse_pos))
+            {
+                target.draw(grid_tile);
+                return;
+            }
+        }
+    }
+}
+
+void draw_shooter_select(sf::RenderTarget &target,
+  const std::vector<shooter_select> &selectors)
+{
+    for (const auto &selector : selectors)
+    {
+        target.draw(selector);
+    }
+}
 
 
 int main(int argc, char *argv[])
@@ -298,10 +396,19 @@ int main(int argc, char *argv[])
         factory.set_enemy_turning_points(
           get_turning_points(map.get_map()->ly_head));
 
+        std::vector<shooter_select> shooter_selectors;
+        shooter_selectors.emplace_back(sf::Color::Cyan,
+          shooter_type::wizard,
+          *manager.get<sf::Texture>("../characters.png"),
+          factory.get_wizard_frames().back());
 
+        Random rd;
         // ecs
         ginseng::database db;
-        factory.spawn_wizard(db);
+        factory.spawn_wizard(db, rd.get_vec2<sf::Vector2f>(100, 500));
+        factory.spawn_wizard(db, rd.get_vec2<sf::Vector2f>(100, 500));
+        factory.spawn_wizard(db, rd.get_vec2<sf::Vector2f>(100, 500));
+        factory.spawn_wizard(db, rd.get_vec2<sf::Vector2f>(100, 500));
 
 
         sf::Font f;
@@ -321,12 +428,19 @@ int main(int argc, char *argv[])
                 {
                     mouse_pos =
                       win.mapPixelToCoords({ ev.mouseMove.x, ev.mouseMove.y });
-                    db.visit([&](components::Animation &anim, components::player_tag) {
-                        anim.setRotation(
-                          std::atan2(mouse_pos.y - anim.getPosition().y,
-                            mouse_pos.x - anim.getPosition().x)
-                          * 180 / 3.14159265);
-                    });
+                    db.visit(
+                      [&](components::Animation &anim, components::player_tag) {
+                          // rotate around its origin
+                          auto direction =
+                            normalize(mouse_pos - anim.getPosition());
+                          if (direction.x < 0)
+                          {
+                              flip_sprite(anim, true, false);
+                          } else
+                          {
+                              flip_sprite(anim, false, false);
+                          }
+                      });
                 }
                 if (ev.type == sf::Event::KeyPressed)
                 {
@@ -336,27 +450,32 @@ int main(int argc, char *argv[])
                     } else if (ev.key.code == sf::Keyboard::D)
                     {
                         factory.spawn_demon(db);
-                    } else if (ev.key.code == sf::Keyboard::Space)
-                    {
-                        db.visit([&](components::Animation &anim,
-                                   components::player_tag) {
-                            auto bullet_id = db.create_entity();
-                            components::Projectile projectile;
-                            // calculate the direction of the bullet with mouse
-                            // position
-                            projectile.direction =
-                              normalize(mouse_pos - anim.getPosition());
-                            projectile.speed = 500;
-                            projectile.damage = 10;
-                            projectile.setSize({ 5, 5 });
-                            projectile.setPosition(anim.getPosition());
-                            db.add_component(bullet_id, projectile);
-                        });
                     }
+                }
+                if (ev.type == sf::Event::MouseButtonPressed
+                    && ev.mouseButton.button == sf::Mouse::Left)
+                {
+                    db.visit(
+                      [&](components::Animation &anim, components::player_tag) {
+                          auto bullet_id = db.create_entity();
+                          components::Projectile projectile;
+                          // calculate the direction of the bullet with mouse
+                          // position
+                          projectile.direction =
+                            normalize(mouse_pos - anim.getPosition());
+                          projectile.speed = 500;
+                          projectile.damage = 5;
+                          projectile.setSize({ 10, 10 });
+                          // rotate the bullet by a fixed amount
+                          projectile.setPosition(anim.getPosition());
+                          projectile.setOrigin(sprite_center(projectile));
+                          db.add_component(bullet_id, projectile);
+                      });
                 }
             }
             delta = clock.restart().asSeconds();
             db.visit([&](components::Animation &anim) { anim.update(delta); });
+
             db.visit([&](ginseng::database::ent_id id,
                        components::Animation &anim,
                        components::EnemyPath &path,
@@ -366,33 +485,47 @@ int main(int argc, char *argv[])
                       factory.get_enemy_turning_points(),
                       delta,
                       path,
-                      vel))
+                      vel)
+                    || health.health <= 0)
                 {
                     db.destroy_entity(id);
                 }
-
-                if (health.health <= 0) { db.destroy_entity(id); }
             });
             db.visit([&](ginseng::database::ent_id id,
                        components::Projectile &projectile) {
-                projectile.move(projectile.direction * projectile.speed * delta);
-                db.visit([&](components::Animation &anim,
-                           components::Health &health) {
-                    if (projectile.getGlobalBounds().intersects(anim.getGlobalBounds()))
-                    {
-                        health.health -= projectile.damage;
-                        db.destroy_entity(id);
-                    }
-                });
+                if (projectile.getPosition().x < 0
+                    || projectile.getPosition().x > win_size.x
+                    || projectile.getPosition().y < 0
+                    || projectile.getPosition().y > win_size.y)
+                {
+                    db.destroy_entity(id);
+                } else
+                {
+                    projectile.move(
+                      projectile.direction * projectile.speed * delta);
+                    projectile.rotate(delta * 1000);
+
+                    db.visit([&](components::Animation &anim,
+                               components::Health &health) {
+                        if (db.exists(id)
+                            && projectile.getGlobalBounds().intersects(
+                              anim.getGlobalBounds()))
+                        {
+                            health.health -= projectile.damage;
+                            db.destroy_entity(id);
+                        }
+                    });
+                }
             });
 
 
             win.clear();
             win.draw(map);
-            db.visit([&](Animation &anim,
-                       ginseng::optional<components::Health> health) {
-                win.draw(anim);
-            });
+            db.visit(
+              [&](Animation &anim,
+                ginseng::optional<components::draw_bounding_box> box_tag) {
+                  win.draw(anim);
+              });
             db.visit([&](Animation &anim, components::Health health) {
                 sf::RectangleShape outsiderect;
                 // put it above the sprite of the width of the sprite and a
@@ -415,9 +548,10 @@ int main(int argc, char *argv[])
             db.visit([&](components::Projectile &projectile) {
                 win.draw(projectile);
             });
+            draw_shooter_select(win, shooter_selectors);
+            draw_select_tile(win, mouse_pos);
 
             win.display();
-
         }
     } catch (const std::exception &e)
     {
